requirements.txt
arrow==0.5.4
bcrypt==1.1.1
cffi==0.9.2
Flask==0.10.1
Flask-Admin==1.1.0
Flask-Bcrypt==0.6.2
Flask-Login==0.2.11
flask-mongoengine==0.7.1
Flask-Script==2.0.5
Flask-SocketIO==0.6.0
Flask-WTF==0.11
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.6
html5lib==0.999
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.4
MarkupSafe==0.23
mongoengine==0.9.0
MySQL-python==1.2.5
phpserialize==1.3
Pillow==2.8.1
pycparser==2.12
pyjade==3.0.0
pymongo==2.8
python-bcrypt==0.3.1
python-dateutil==2.4.2
python-slugify==1.0.2
six==1.9.0
Unidecode==0.4.17
Werkzeug==0.10.4
WTForms==2.0.2
